requirements.txt
fastapi
uvicorn
httpx
jinja2
pydantic
pydantic-settings
brotli
aiohttp
pyyaml